Create error code lookup function
Reduce file name from error_code_registry to errors. Use lookup function to access the error codes and throw a ValueError in case the error could not be found. Add all error hexa codes to the registry. The additional data is not replacing the error description, so remove test covering that incorrect case.