Add '@=' separator to pass JSON objects via CLI
The current '=' separator enables to add key-value pairs as strings
in a JSON object to the HTTP body. But there's no way to provide
more complex JSON objects

This commit introduces a new '@=' that adds a key-value pair where
the right value is evaluated to a JSON object, allowing more complex
structures in the HTTP body.

Change-Id: I126a6c8529a6dbf9108b904b2b4d93ef60309113